requirements.txt
# In requirements.txt

# --- Core ML & Computer Vision ---
ultralytics
opencv-python

# --- Data Handling & Plotting ---
pandas
matplotlib
PyYAML
tqdm

# --- Data Downloading ---
synapseclient
requests
python-dotenv